updated backend, fixed delete functionality, added dynamic routing
